requirements.txt
fastapi==0.104.1
pydantic==2.5.2
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
imbalanced-learn==0.12.3
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2